+++ docs/structs.md
@@ -2,7 +2,7 @@
 
 Structs in Kiwi provide a way to bundle data and functionality together.
 
-### Defining Structs
+## Defining Structs
 
 To define a struct in Kiwi, use the `struct` keyword followed by the struct name and a block of code defining its properties and methods.
 
@@ -16,12 +16,12 @@ struct MyStruct
   end
 
   fn say_hello()
-    println("Hello, ${name}!")
+    println("Hello, ${@name}!")
   end
 end
 ```
 
-### Creating Instances
+## Creating Instances
 
 To create an instance of a struct, use the `.new()` method followed by any arguments the constructor accepts.
 
@@ -30,19 +30,19 @@ my_object = MyStruct.new("Kiwi")
 my_object.say_hello()  # prints: Hello, Kiwi!
 ```
 
-### Inheritance
+## Inheritance
 
 Kiwi supports single inheritance. Use the `<` symbol to specify the parent struct.
 
 ```kiwi
 struct MySubStruct < MyStruct
   fn say_goodbye()
-    println("Goodbye, ${name}!")
+    println("Goodbye, ${@name}!")
   end
 end
 ```
 
-### Method Definition
+## Method Definition
 
 Methods are defined using the `fn` keyword, followed by the method name and any parameters. Use `@` to access the current instance.
 
@@ -54,7 +54,7 @@ struct MyStruct
 end
 ```
 
-### Overriding `to_string()`
+## Overriding `to_string()`
 
 The `override` keyword is not required to override `to_string()`.
 
@@ -71,7 +71,7 @@ println(instance)    # prints: I am a Kiwi struct
 println(string_repr) # prints: I am a Kiwi struct
 ```
 
-### Static Method Definition
+## Static Method Definition
 
 Methods declared as static can be invoked directly through the struct and cannot be invoked through an instance.
 
@@ -85,7 +85,7 @@ end
 MyStruct.static_method() # prints: I can be invoked without an instance!
 ```
 
-### Access Control
+## Access Control
 
 Kiwi supports `private` methods that cannot be called outside the struct definition.
 
@@ -97,7 +97,7 @@ struct MyStruct
 end
 ```
 
-All instance variables are private by default. You must explicitly implement accessors and modifiers to access them outside of the struct.
+All instance variables are public but readonly. You can explicitly implement accessors and modifiers to control access to them outside of struct scope.
 
 ```kiwi
 struct MyStruct
@@ -107,4 +107,59 @@ struct MyStruct
 end
 
 inst = MyStruct.new("kiwi")
+println inst.name # kiwi
+```
+
+## Type-checking, member queries
+
+```kiwi
+struct Hashable
+  fn to_hash()
+    {}
+  end
+end
+
+struct A < Hashable
+  fn new(name)
+    @name = name
+  end
+
+  fn to_hash()
+    { name: @name }
+  end
+
+  fn say_hello()
+    println("Hello, ${@name}!")
+  end
+end
+
+struct B < A
+  fn say_goodbye()
+    println("Goodbye, ${@name}!")
+    foobar()
+  end
+
+  private fn foobar()
+    println "foobar is private"
+  end
+end
+
+my_obj = B.new("Kiwi")
+println my_obj.to_hash()      # { "name": "Kiwi" }
+
+# type checking
+println my_obj.is_a(Hashable) # true
+
+# set a member
+my_obj.set("name", "foobar")
+
+# query members
+println my_obj.keys()         # ["name"]
+println my_obj.values()       # ["foobar"]
+
+# clone an object
+x = my_obj.clone()
+println x                     # <struct(name=B, identifier=x>
+
+x.say_goodbye()               # Goodbye, foobar!\nfoobar is private
 ```
